Unbitrot and fix memory corruption issues
MoarVM internals have changed quite a bit since this code was written, so adapt
the code to what is now needed. This replaces a few hard coded assumptions with
values introspected from the compiler's configuration. Since MoarVM's dispatch
mechanism got replaced in a recent version, we need to switch as well, which
means that we now require a very recent Rakudo.

This also fixes memory corruption issues which resulted from us giving pointers
to stack variables to MoarVM which would access them long after we exited from
the frame they were valid in.